feat: move environment into modules
