+++ libs/@local/hashql/core/src/type/environment/variance.rs
@@ -1,0 +1,32 @@
+/// Represents the type relationship variance used in generic type checking.
+///
+/// Variance defines how subtyping relationships between generic parameter types
+/// affect subtyping relationships between the resulting parameterized types.
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Default)]
+pub enum Variance {
+    /// Covariance
+    ///
+    /// Covariance preserves the ordering of types, suppose `A` and `B` are types, and `I` is a type
+    /// constructor. If `I` is covariant, and `A <: B` (A is a subtype of B), then `I<A> <: I<B>`.
+    ///
+    /// Example: If `Cat <: Animal`, then `List<Cat> <: List<Animal>` in a covariant position.
+    #[default]
+    Covariant,
+    /// Contravariance
+    ///
+    /// Contravariance reverses the ordering of types, suppose `A` and `B` are types, and `I` is a
+    /// type constructor. If `I` is contravariant, and `A <: B` (A is a subtype of B), then
+    /// `I<B> <: I<A>`.
+    ///
+    /// Example: If `Cat <: Animal`, then `fn(Animal) -> Null <: fn(Cat) -> Void`
+    /// for parameter types in a contravariant position.
+    Contravariant,
+    /// Invariance
+    ///
+    /// Invariance requires exact match of types, suppose `A` and `B` are types, and `I` is a type
+    /// constructor. If `I` is invariant, then `I<A> <: I<B>` if and only if `A = B`.
+    ///
+    /// Example: `Dict<Integer, String> <: Dict<Number, String>` is not a subtype
+    /// of the other, even though `Integer <: Number`.
+    Invariant,
+}
